+++ vulture-whitelist.py
@@ -260,9 +260,9 @@ animation_duration  # unused variable (src/immich_memories/titles/renderer_pil.p
 _._title_font  # unused attribute (src/immich_memories/titles/renderer_pil.py:92)
 _._subtitle_font  # unused attribute (src/immich_memories/titles/renderer_pil.py:93)
 _.render_all_frames_parallel  # unused method (src/immich_memories/titles/renderer_pil.py:265)
-birthday_particle_count  # unused variable (src/immich_memories/titles/renderer_taichi.py:91)
-fireworks_fade_speed  # unused variable (src/immich_memories/titles/renderer_taichi.py:99)
-split_title_lines  # unused function (src/immich_memories/titles/renderer_taichi.py:169)
+birthday_particle_count  # unused variable (src/immich_memories/titles/renderer_taichi.py:82)
+fireworks_fade_speed  # unused variable (src/immich_memories/titles/renderer_taichi.py:90)
+split_title_lines  # unused function (src/immich_memories/titles/taichi_text.py:69)
 render_text_sdf  # unused function (src/immich_memories/titles/sdf_font_rendering.py:199)
 measure_text  # unused function (src/immich_memories/titles/sdf_font_rendering.py:248)
 letter_spacing  # unused variable (src/immich_memories/titles/styles.py:27)
